unahealth/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
et-xmlfile==1.1.0
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
